backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
python-dotenv==1.2.1
google-generativeai==0.8.3
pydantic==2.12.3
httpx==0.28.1
tweepy==4.15.0
pytest==8.3.5
pytest-asyncio==0.25.3
